requirements.txt
python-dateutil==2.4.2
w1thermsensor==0.2.2
boto3==1.2.2
